fixed the bugs in the extract_dataframe.py
