+++ .gitignore
@@ -1,2 +1,4 @@
 __pycache__/
 data/
+africa_twitter_data.json
+global_twitter_data.json
